Fix wrong handling of region if number of characters changes
For example suppose we have the following where | is point and $ is mark and
execute evil-numbers/inc-at-point.

```
|9 9$
```

Expected:

```
10 10
```

Actual:

```
10 9
```

This is because we would save the beginning and end of the region before we do
any changes, so the second 9 is pushed out of the saved region bounds by the
first 9 turning into a 10.