fix: handle hyphen-separated numbers correctly
Imagine we're editing hyphen-separated numbers like the following:

    cursor:          v
    text:   2020-01-02

When we invoke 'inc-at-pt', we want the last number to change:

    cursor:          v
    text:   2020-01-01

(Note that the number is decremented instead of being incremented due to
the leading '-'. This also matches Vim behavior)

However, evil-numbers currently change the first number instead:

    cursor:    v
    text:   2021-01-02

This patch fixes evil-numbers to match the expected behavior.